pyproject.toml
[project]
name = "pytemplate"
version = "0.1.0"
description = "Python package template with Poetry"

[[project.authors]]
name = "kezure <3447723+kezure@users.noreply.github.com>"

[tool.poetry]
name = "pytemplate"
version = "0.1.0"
description = "Python package template with Poetry"
authors = ["kezure <3447723+kezure@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.3.0"
ruff = "^0.14.0"
mypy = "^1.18.2"
pytest = "^8.4.2"
pytest-xdist = "^3.8.0"
pytest-cov = "^7.0.0"
python-dotenv = "^1.1.1"
tbvaccine = "^0.3.1"
sphinx = "^8.2.3"
sphinx-autodoc-typehints = "^3.5.1"
sphinx-rtd-theme = "^3.0.2"
sphinx-pyproject = "^0.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpaths = ["pytemplate"]
testpaths = ["tests"]
addopts = [
    "-s",
    "--durations=0",
    # "-vv",
]

[tool.coverage.run]
omit = [
    "/tmp/*",
    "tests/conftest.py"
]
concurrency = ["multiprocessing"]
parallel = "true"

[tool.sphinx-pyproject]
project = "pytemplate"
copyright = "2023, 3447723+kezure@users.noreply.github.com"
language = "ja"
package_root = "pytemplate"
html_theme = "sphinx_rtd_theme"
todo_include_todos = true
templates_path = ["_templates"]
html_static_path = ["_static"]
exclude_patterns = ["_build", "Thumbs.db", ".DS_Store", "tests"]
extensions = [
    "sphinx.ext.autodoc",
    "sphinx_autodoc_typehints",
    "sphinx.ext.autosummary",
    "sphinx.ext.napoleon",
    "sphinx.ext.viewcode",
    "sphinx.ext.intersphinx",
    "sphinx.ext.todo",
    "sphinx.ext.githubpages",
]

[tool.ruff]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = ["ALL"]
ignore = [
#   "E203", # Whitespace before ':' (E203)
  "E501", # Line too long (82 &gt; 79 characters) (E501)
#   "F811", # Redefinition of unused name from line n (F811)
#   "E741", # Do not define classes named 'I', 'O', or 'l' (E742)
#   "E266", # too many leading '#' for block comment (E266)
#   "PIE",  # flake8-pie (PIE)
#   "F601", # multi-value-repeated-key-literal (F601)
#   "E721", # type-comparison (E721)
    "LOG015", # `info()` call on root logger (LOG015)
]
extend-ignore = [
    
]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.ruff.per-file-ignores]
"__init__.py" = ["D104"]
